pyproject.toml
[tool.poetry]
name = "support-tg-bot"
version = "0.1.0"
description = ""
authors = ["falcolnic <mykh.yano@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.6"
django-environ = "^0.11.2"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
dishka = "^1.4.0"
faststream = {extras = ["cli"], version = "^0.5.29"}
aiokafka = "^0.12.0"
watchfiles = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
